backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.11.1
pytest==8.3.5
httpx==0.28.1
python-dotenv==1.0.1

jsonschema==4.23.0
